+++ samples/subsys/dap/prj.conf
@@ -1,19 +1,18 @@
+CONFIG_USB_DEVICE_STACK_NEXT=y
+CONFIG_SAMPLE_USBD_PRODUCT="Zephyr CMSIS-DAP"
+CONFIG_SAMPLE_USBD_PID=0x0204
+CONFIG_SAMPLE_USBD_20_EXTENSION_DESC=y
+
 CONFIG_STDOUT_CONSOLE=y
-CONFIG_USB_DEVICE_STACK=y
-CONFIG_USB_DEVICE_BOS=y
 CONFIG_SERIAL=y
-CONFIG_USB_DEVICE_PRODUCT="Zephyr CMSIS-DAP"
-CONFIG_USB_DEVICE_PID=0x0204
 CONFIG_UART_INTERRUPT_DRIVEN=y
 CONFIG_UART_LINE_CTRL=y
-CONFIG_USB_DEVICE_INITIALIZE_AT_BOOT=n
 
 CONFIG_LOG=y
+CONFIG_LOG_DEFAULT_LEVEL=2
 
-CONFIG_USB_DRIVER_LOG_LEVEL_INF=y
-CONFIG_USB_DEVICE_LOG_LEVEL_INF=y
 CONFIG_DAP_LOG_LEVEL_INF=y
 CONFIG_DP_DRIVER_LOG_LEVEL_INF=y
 CONFIG_DAP=y
+CONFIG_DAP_BACKEND_USB=y
 CONFIG_GPIO=y
-CONFIG_NET_BUF=y
